Unsupervised Verification/main.py   SVM outvalues design UNDER DEVELOP
Implemented global variables:
 outlier_values
 inlier_values

Finding the values of variable via"
     outlier_values = np.array(pd.DataFrame(data).iloc[outlier_index])     74row
    inlier_index = np.array(pd.DataFrame(data).iloc[inlier_index])            75row

printed to the separated .csv

Here is needed to generated ones and zeroes arrays and concetenate with outlier, inlier together. The values will dependent on length of outlier and inlier. After that the sum of outlier and inlier have to be same with length of frame length